enables async login
login is running successfully, using an async message stream to feed
events to godot. It's still not working perfectly yet. The server says
the login is successful, but I'm not able to push a client update from
the success of a new session. this might be due to the way session is
working, and it never actually returning? It seems unlikely with the way
my tests were written, but we'll see.
I would like to implement a nice loading bar animation for the login