adds scene switching
allows scenes to be switced, using the session scene as a root.
this makes sure that the update checking continues through scenes.
unfortunately the login is still getting stuck waiting for the
metaverse_session to initialize. The loading screen does not actually work
at the moment, because of the way the scene's onready works. The login
tells the session to load in a new scene, which means it runs its _ready
function to get to the ready state. trying to run it after the scene has
loaded is very hard to figure out.